add(): implement path exec

+++ hsh_exec_cmd.c
@@ -8,14 +8,29 @@
 
 void hsh_exec_cmd(char **arguments)
 {
+	int i = 0;
+	struct stat st;
+	char extern **environ;
+	char **arr = array_PATH(environ, arguments);
+
 	if (arguments[0] != NULL)
 	{
 		if (execve(arguments[0], arguments, NULL) == -1)
 		{
-			perror("Execution Error\n");
-			kill(getpid(), SIGTERM);
+			while (arr[i])
+			{
+				if (stat(arr[i], &st) == 0)
+				{
+					arguments[0] = arr[i];
+					if (execve(arguments[0], arguments, NULL) == -1)
+					{
+						perror("Execution Error\n");
+						kill(getpid(), SIGTERM);
+					}
+				}
+				i++;
+			}
 		}
-	
 	}
 	kill(getpid(), SIGTERM);
 }
